backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.46.0",
]
